+++ examples/resources/automq_environment/provider_stub.tf
@@ -1,0 +1,4 @@
+provider "automq" {
+  automq_byoc_access_key_id = "access-key"
+  automq_byoc_secret_key    = "secret-key"
+}